refactor(frontend): extract duplicated timeline item logic to AppointmentTimelineItem
